Support for AspectJ parser

+++ platform/lang-api/src/com/intellij/lang/PsiBuilderUtil.java
@@ -114,7 +114,7 @@ public class PsiBuilderUtil {
    */
   public static void drop(final PsiBuilder.Marker... markers) {
     for (PsiBuilder.Marker marker : markers) {
-      marker.drop();
+      if (marker != null) marker.drop();
     }
   }
 }
